Add delete functionality

+++ aws_mfa_config/profile.py
@@ -16,7 +16,7 @@ def get_input(prompt, default_value, mask=False):
     if mask:
         existing_value = mask_value(default_value)
 
-    return input('{} [{}]: '.format(prompt, existing_value)).strip or default_value
+    return input('{} [{}]: '.format(prompt, existing_value)).strip() or default_value
 
 def add(context):
     """ Parse the command line arguments and return a context object """
@@ -76,6 +76,50 @@ def add(context):
 
     if aws_mfa_device_arn != '':
         credentials[profile_name]['aws_mfa_device_arn'] = aws_mfa_device_arn
+
+    with open(config_path, 'w') as configfile:
+        config.write(configfile)
+
+    with open(credentials_path, 'w') as credentialsfile:
+        credentials.write(credentialsfile)
+
+def delete(context):
+    """ Delete an AWS Profile from the config files"""
+
+    credentials_path = os.path.join(context.config_path, 'credentials')
+    config_path = os.path.join(context.config_path, 'config')
+
+    if ((not os.path.exists(credentials_path)) or (not os.path.exists(config_path))):
+        print('Could not locate a credentials or a config file in \
+		{}.'.format(context.config_path))
+
+        sys.exit()
+
+    profile_name = context.profile_name
+
+    choice = input('Are you sure you want to delete the profile with name {}? [y/N] '.format(
+        profile_name)).lower()
+
+    if choice != 'y':
+        sys.exit()
+
+    config = configparser.ConfigParser()
+    credentials = configparser.ConfigParser()
+    config.read('/'.join([context.config_path, 'config']))
+    credentials.read('/'.join([context.config_path, 'credentials']))
+
+    config.remove_section(profile_name)
+    credentials.remove_section(profile_name)
+
+    mfa_profile_name = '{}-mfa'.format(profile_name)
+    if config.has_section(mfa_profile_name) or credentials.has_section(mfa_profile_name):
+        choice = input('An MFA profile was found ({}). Do you want to delete it as well? [Y/n] '.format(
+            mfa_profile_name)).lower()
+        if choice == 'n':
+            sys.exit()
+
+        config.remove_section(mfa_profile_name)
+        credentials.remove_section(mfa_profile_name)
 
     with open(config_path, 'w') as configfile:
         config.write(configfile)
